# specs/001-personal-finance-manager/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Estateably Finance API
  version: 1.1.0
  description: |
    Personal finance manager — single user, single currency (USD), double-entry ledger.

    This document is the **frozen contract** (constitution Principle IV) and the complete API
    documentation. Served verbatim at `/docs` by the web app. Any change to it goes back
    through the plan.

    **Money** travels as JSON **strings of integer cents** (`"4250"` = $42.50), never as JSON
    numbers: `JSON.parse` reads a JSON number as a float and can corrupt large values before
    validation runs. Arithmetic uses `bigint`. Amounts are bounded to ±10^15 cents — a
    technical guard protecting the database's BIGINT columns, validated after `BigInt`
    conversion, rejected with `VALIDATION_FAILED`.

    **Dates** are calendar dates, `YYYY-MM-DD`, no time component. "Today" is the current
    date in the `APP_TIMEZONE` environment variable's timezone.

    **Ids** are serialized as decimal strings (they are BIGINT).

    **Errors** all share one shape (`ErrorResponse`) with a machine-readable `code`, a
    human-readable `message`, optional field-level `details`, and the request's
    `correlationId`. Every response (success or error) also carries the correlation id in the
    `X-Correlation-Id` header. Referencing a non-existent or archived entity in a request
    *body* is a domain rule violation (`422 DOMAIN_RULE_VIOLATION`); `404 NOT_FOUND` is
    reserved for the id in the URL path.

...
servers:
...
paths:
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List accounts with current balances
      description: >
        Returns every account (archived ones only when `includeArchived=true`), each with its
        current balance, plus the total across non-archived accounts. Accounts are few, so
        this list is not paginated.
      parameters:
        - name: includeArchived
          in: query
          required: false
          description: Include archived accounts in the list. Defaults to false.
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Accounts and the non-archived total.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountListResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      description: >
        Creates an account. A non-zero `openingBalance` records an *opening* transaction dated
        `openingDate` through the ledger; a zero one records nothing. The opening balance is
        signed and may be negative (a card that already owes money).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountCreateRequest' }
      responses:
        '201':
          description: The created account, balance included.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/DuplicateName' }
        '500': { $ref: '#/components/responses/InternalError' }
  /accounts/balance-history:
    get:
      operationId: getBalanceHistory
      tags: [accounts]
      summary: Daily balances over a date range
      description: >
        One balance per calendar day from `from` to `to` inclusive, for each account and for the
        total. Element `i` of every `balances` array is the balance at the end of day
        `from + i`: the sum of the account's non-deleted entries dated on or before that day,
        the same value `getAccountBalanceAsOf` answers. `total` sums the non-archived accounts
        only, whether or not `includeArchived` is set. Read-only; derived from entries on every
        call, never stored.
      parameters:
        - name: from
          in: query
          required: false
          description: >
            First day, inclusive. When absent, the earliest entry date of the non-archived
            accounts, whether or not `includeArchived` is set, clamped to 10 years before `to`;
            `to` itself when they have no entries. Must not be after `to`, and the span must not
            exceed 10 years.
          schema: { type: string, format: date }
          example: '2026-08-24'
        - name: to
          in: query
          required: false
          description: >
            Last day, inclusive. Defaults to today in `APP_TIMEZONE`. Must not be after today,
            otherwise `VALIDATION_FAILED`.
          schema: { type: string, format: date }
          example: '2026-09-22'
        - name: includeArchived
          in: query
          required: false
          description: Also return archived accounts' series. They never count toward `total`.
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: The daily series.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BalanceHistoryResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalError' }
  /accounts/{id}:
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Edit an account
      description: >
        Renames the account or changes its kind, opening balance or opening date. Opening
        balance and date edit the account's *opening* transaction through the ledger: zero to
        non-zero creates it, non-zero to zero deletes it, any other change rebalances it.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountUpdateRequest' }
      responses:
        '200':
          description: The updated account.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/DuplicateName' }
        '422': { $ref: '#/components/responses/DomainRuleViolation' }
        '500': { $ref: '#/components/responses/InternalError' }
  /accounts/{id}/archive:
    post:
      operationId: archiveAccount
      tags: [accounts]
      summary: Archive an account
      description: >
        Hides the account from pickers and dashboard totals. Its history stays readable and no
        transaction is altered. Reversible with `unarchiveAccount`.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: The archived account.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /accounts/{id}/unarchive:
    post:
      operationId: unarchiveAccount
      tags: [accounts]
      summary: Unarchive an account
      description: Restores the account to pickers and totals with its balance and history intact.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: The unarchived account.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /accounts/{id}/balance:
    get:
      operationId: getAccountBalanceAsOf
      tags: [accounts]
      summary: Balance of an account as of a date
      description: >
        The sum of the account's non-deleted entries dated on or before `asOf`. A date before
        the first transaction answers `"0"`. Correct on every read, cached values present or not.
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: asOf
          in: query
          required: true
          description: The date to compute the balance at, inclusive.
          schema: { type: string, format: date }
          example: '2026-09-12'
      responses:
        '200':
          description: The balance at that date.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BalanceResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      description: >
        Lists transactions globally or filtered by account, kind, category, project and date
        range, ordered by `date` descending, ties broken by `id` descending (stable across
        pages). Paginated by `limit`/`offset`; the response carries the total match count. An
        offset past the last row returns an empty page with the correct total. A date range
        must start on or before its end and span at most 24 months.
      parameters:
        - name: accountId
          in: query
          required: false
          description: Only transactions touching this account.
          schema: { $ref: '#/components/schemas/Id' }
        - name: kind
          in: query
          required: false
          description: Only this kind.
          schema: { $ref: '#/components/schemas/TransactionKind' }
        - name: categoryId
          in: query
          required: false
          description: Only this category.
          schema: { $ref: '#/components/schemas/Id' }
        - name: projectId
          in: query
          required: false
          description: Only this project.
          schema: { $ref: '#/components/schemas/Id' }
        - $ref: '#/components/parameters/FromParam'
        - $ref: '#/components/parameters/ToParam'
        - $ref: '#/components/parameters/QueryParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: One page of transactions plus the total match count.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedTransactions' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Record a transaction
      description: >
        Records an expense, income or transfer (an *opening* transaction is created only
        through the account endpoints). The intent is translated into exactly two balanced
        ledger entries in one atomic write. Rejected with a structured error: non-positive
        amounts, future dates, archived or non-existent accounts/categories, category types
        that do not match the kind, transfers between the same account, projects on anything
        but an expense, closed projects.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionRequest' }
      responses:
        '201':
          description: The recorded transaction.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '422': { $ref: '#/components/responses/DomainRuleViolation' }
        '500': { $ref: '#/components/responses/InternalError' }
  /transactions/{id}:
    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Edit a transaction
      description: >
        Replaces the transaction with a complete new intent — kind included, so an expense may
        become an income or a transfer and back — validated exactly like a newly recorded one.
        All entries are regenerated in one atomic write. An *opening* transaction is not
        editable here (it stays *opening*, edited through the account form).
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionRequest' }
      responses:
        '200':
          description: The updated transaction.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/DomainRuleViolation' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      description: >
        Soft-deletes the transaction: it disappears from every list, balance and report, and
        the ledger still sums to zero. An *opening* transaction is deleted by setting its
        account's opening balance to zero, not here.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Deleted; no content.
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/DomainRuleViolation' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: The entity's id.
      schema: { $ref: '#/components/schemas/Id' }
    LimitParam:
      name: limit
      in: query
      required: false
      description: Page size, 1–200. Above 200 is rejected with `VALIDATION_FAILED`.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    OffsetParam:
      name: offset
      in: query
      required: false
      description: Rows to skip. An offset past the last row returns an empty page with the correct total.
      schema: { type: integer, minimum: 0, default: 0 }
    FromParam:
      name: from
      in: query
      required: false
      description: Range start, inclusive. With `to`, must not start after it nor span more than 24 months.
      schema: { type: string, format: date }
      example: '2026-09-01'
    ToParam:
      name: to
      in: query
      required: false
      description: Range end, inclusive.
      schema: { type: string, format: date }
      example: '2026-09-30'
    QueryParam:
      name: q
      in: query
      required: false
      description: >
        Only transactions whose description contains this text, ignoring case. Matched
        literally: `%`, `_` and `\` are not wildcards. Trimmed; 1–60 characters after trimming,
        otherwise `VALIDATION_FAILED`. Combined with the other filters by AND. Order and
        pagination are unchanged.
      schema: { type: string, minLength: 1, maxLength: 60 }
      example: uber
...
  headers:
    XCorrelationId:
      description: The request's correlation id; matches the `correlationId` in the body and the stdout log line.
      schema: { type: string, format: uuid }
  responses:
    ValidationError:
      description: The request's shape, a field's format, the money guard, a date rule or a pagination bound failed (`VALIDATION_FAILED`).
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/XCorrelationId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: VALIDATION_FAILED
            message: Validation failed
            details:
              [
                {
                  field: amount,
                  message: 'amount must be a string of integer cents with absolute value at most 10^15',
                },
              ]
            correlationId: 6f1b0c1e-8a24-4a5f-9b6d-2f3a7c1d9e10
    NotFound:
      description: The id in the URL path does not exist or is deleted (`NOT_FOUND`).
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/XCorrelationId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: NOT_FOUND
            message: Transaction 42 does not exist
            correlationId: 6f1b0c1e-8a24-4a5f-9b6d-2f3a7c1d9e10
    DuplicateName:
      description: Another account, category or project (archived and closed rows included) already holds this name, compared case- and whitespace-insensitively (`DUPLICATE_NAME`).
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/XCorrelationId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: DUPLICATE_NAME
            message: An account named "Checking" already exists
            details: [{ field: name, message: 'name is already in use' }]
            correlationId: 6f1b0c1e-8a24-4a5f-9b6d-2f3a7c1d9e10
    DomainRuleViolation:
      description: >
        A domain rule rejected the request (`DOMAIN_RULE_VIOLATION`): archived account or
        category, category type mismatch, transfer between the same account, closed or
        non-existent referenced entity, project on a non-expense, delete of a used project,
        end date before due date, and similar.
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/XCorrelationId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: DOMAIN_RULE_VIOLATION
            message: A transfer cannot use the same account as source and destination
            details: [{ field: counterAccountId, message: 'must differ from accountId' }]
            correlationId: 6f1b0c1e-8a24-4a5f-9b6d-2f3a7c1d9e10
...
    InternalError:
      description: Unhandled error (`INTERNAL_ERROR`); logged with stack trace under the same correlation id.
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/XCorrelationId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: INTERNAL_ERROR
            message: Something went wrong
            correlationId: 6f1b0c1e-8a24-4a5f-9b6d-2f3a7c1d9e10
  schemas:
    Id:
      type: string
      pattern: '^[0-9]+$'
      description: Entity id, a BIGINT serialized as a decimal string.
      example: '42'
    Money:
      type: string
      pattern: '^-?[0-9]{1,16}$'
      description: >
        Integer cents as a decimal string; never a JSON number. The ±10^15-cent guard is
        validated after bigint conversion, not by this pattern.
      example: '4250'
    DateOnly:
      type: string
      format: date
      description: Calendar date, `YYYY-MM-DD`, no time component.
      example: '2026-09-10'
    AccountKind:
      type: string
      enum: [bank, cash, card]
      description: The user-facing kind of an account.
      example: bank
...
    TransactionKind:
      type: string
      enum: [expense, income, transfer, opening]
      description: >
        The intent a transaction expresses. `opening` appears only in responses; it is created
        and edited through the account endpoints.
      example: expense
...
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - NOT_FOUND
        - DUPLICATE_NAME
        - DOMAIN_RULE_VIOLATION
        - AI_NOT_CONFIGURED
        - AI_PROVIDER_ERROR
        - AI_RATE_LIMITED
        - AI_TIMEOUT
        - INTERNAL_ERROR
      description: Every error code the API can answer with. One code per failure class.
...
    ErrorResponse:
      type: object
      required: [code, message, correlationId]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message:
          type: string
          description: Human-readable summary, safe to show the user.
          example: Validation failed
        details:
          type: array
          items: { $ref: '#/components/schemas/ErrorDetail' }
          description: Field-level details when a field is nameable; absent otherwise.
        correlationId:
          type: string
          format: uuid
          description: Matches the `X-Correlation-Id` header and exactly one stdout log line.
          example: 6f1b0c1e-8a24-4a5f-9b6d-2f3a7c1d9e10
    AccountCreateRequest:
      type: object
      required: [name, kind, openingBalance, openingDate]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
          description: 1–60 characters after trimming; unique among accounts, archived included.
          example: Checking
        kind: { $ref: '#/components/schemas/AccountKind' }
        openingBalance:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: Signed cents; may be negative (existing debt) or `"0"` (no opening transaction).
          example: '150000'
        openingDate:
          allOf: [{ $ref: '#/components/schemas/DateOnly' }]
          description: The opening transaction's date; not in the future.
          example: '2026-09-01'
    AccountUpdateRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
          example: Everyday Checking
        kind: { $ref: '#/components/schemas/AccountKind' }
        openingBalance:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: New signed opening balance; `"0"` deletes the opening transaction.
          example: '175000'
        openingDate: { $ref: '#/components/schemas/DateOnly' }
    AccountResponse:
      type: object
      required: [id, name, kind, openingBalance, openingDate, archived, balance]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        name: { type: string, example: Checking }
        kind: { $ref: '#/components/schemas/AccountKind' }
        openingBalance:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: Signed cents of the opening transaction; `"0"` when there is none.
          example: '150000'
        openingDate: { $ref: '#/components/schemas/DateOnly' }
        archived: { type: boolean, example: false }
        balance:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: Current balance, signed cents.
          example: '445750'
    AccountListResponse:
      type: object
      required: [items, totalBalance]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AccountResponse' }
        totalBalance:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: Sum of current balances across non-archived accounts, signed cents.
          example: '395750'
    BalanceResponse:
      type: object
      required: [accountId, asOf, balance]
      properties:
        accountId: { $ref: '#/components/schemas/Id' }
        asOf: { $ref: '#/components/schemas/DateOnly' }
        balance:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: Sum of the account's non-deleted entries dated on or before `asOf`.
          example: '145750'
    AccountBalanceSeries:
      type: object
      required: [accountId, archived, balances]
      properties:
        accountId: { $ref: '#/components/schemas/Id' }
        archived: { type: boolean, example: false }
        balances:
          type: array
          items: { $ref: '#/components/schemas/Money' }
          description: One signed-cents balance per day, `from` first.
          example: ['445750', '441500', '441500']
    BalanceHistoryResponse:
      type: object
      required: [from, to, total, accounts]
      properties:
        from: { $ref: '#/components/schemas/DateOnly' }
        to: { $ref: '#/components/schemas/DateOnly' }
        total:
          type: array
          items: { $ref: '#/components/schemas/Money' }
          description: >
            Per day, the sum of the non-archived accounts' balances, signed cents. Its last
            element equals `AccountListResponse.totalBalance` when `to` is today.
          example: ['395750', '391500', '391500']
        accounts:
          type: array
          items: { $ref: '#/components/schemas/AccountBalanceSeries' }
          description: Same accounts and order as `listAccounts` with the same `includeArchived`.
...
    ExpenseIntent:
      type: object
      required: [kind, date, description, amount, accountId, categoryId]
      properties:
        kind: { type: string, enum: [expense], example: expense }
        date:
          allOf: [{ $ref: '#/components/schemas/DateOnly' }]
          description: Not in the future.
        description:
          type: string
          minLength: 1
          maxLength: 120
          description: 1–120 characters after trimming; whitespace-only is rejected.
          example: Market
        amount:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: Positive cents.
          example: '4250'
        accountId:
          allOf: [{ $ref: '#/components/schemas/Id' }]
          description: The non-archived account the money leaves.
        categoryId:
          allOf: [{ $ref: '#/components/schemas/Id' }]
          description: A non-archived category of type `expense`.
        projectId:
          allOf: [{ $ref: '#/components/schemas/Id' }]
          description: Optional; an active project. Only expenses may carry one.
    IncomeIntent:
      type: object
      required: [kind, date, description, amount, accountId, categoryId]
      properties:
        kind: { type: string, enum: [income], example: income }
        date:
          allOf: [{ $ref: '#/components/schemas/DateOnly' }]
          description: Not in the future.
        description:
          type: string
          minLength: 1
          maxLength: 120
          example: September salary
        amount:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: Positive cents.
          example: '300000'
        accountId:
          allOf: [{ $ref: '#/components/schemas/Id' }]
          description: The non-archived account the money enters.
        categoryId:
          allOf: [{ $ref: '#/components/schemas/Id' }]
          description: A non-archived category of type `income`.
    TransferIntent:
      type: object
      required: [kind, date, description, amount, accountId, counterAccountId]
      properties:
        kind: { type: string, enum: [transfer], example: transfer }
        date:
          allOf: [{ $ref: '#/components/schemas/DateOnly' }]
          description: Not in the future.
        description:
          type: string
          minLength: 1
          maxLength: 120
          example: To savings
        amount:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: Positive cents.
          example: '50000'
        accountId:
          allOf: [{ $ref: '#/components/schemas/Id' }]
          description: Source account, non-archived.
        counterAccountId:
          allOf: [{ $ref: '#/components/schemas/Id' }]
          description: Destination account, non-archived, different from `accountId`. Transfers carry no category and no project.
    TransactionRequest:
      oneOf:
        - $ref: '#/components/schemas/ExpenseIntent'
        - $ref: '#/components/schemas/IncomeIntent'
        - $ref: '#/components/schemas/TransferIntent'
      discriminator:
        propertyName: kind
        mapping:
          expense: '#/components/schemas/ExpenseIntent'
          income: '#/components/schemas/IncomeIntent'
          transfer: '#/components/schemas/TransferIntent'
      description: >
        A complete intent, discriminated by `kind`. Used for recording and for editing (an
        edit replaces the whole intent, kind included).
    TransactionResponse:
      type: object
      required: [id, kind, date, description, amount, accountId]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        kind: { $ref: '#/components/schemas/TransactionKind' }
        date: { $ref: '#/components/schemas/DateOnly' }
        description: { type: string, example: Market }
        amount:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: >
            Positive cents for expense, income and transfer; signed cents for `opening`.
          example: '4250'
        accountId:
          allOf: [{ $ref: '#/components/schemas/Id' }]
          description: The account (transfer — the source).
        counterAccountId:
          allOf: [{ $ref: '#/components/schemas/Id' }]
          description: Transfer destination; absent on other kinds.
        categoryId:
          allOf: [{ $ref: '#/components/schemas/Id' }]
          description: Present on expense and income; absent on transfer and opening.
        projectId:
          allOf: [{ $ref: '#/components/schemas/Id' }]
          description: Present when the expense is tagged with a project.
    PaginatedTransactions:
      type: object
      required: [items, total, limit, offset]
      description: The shared paginated shape — `items`, `total`, `limit`, `offset`.
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/TransactionResponse' }
        total:
          type: integer
          minimum: 0
          description: Total matching transactions across all pages.
          example: 87
        limit: { type: integer, example: 50 }
        offset: { type: integer, example: 0 }
...